Fixes platform-specific bug caused by int32 cast

+++ baus/models.py
@@ -337,8 +337,7 @@ def household_relocation(households, household_relocation_rates,
 
     # set households that are moving to a building_id of -1 (means unplaced)
     households.update_col_from_series("building_id",
-                                      pd.Series(-1,
-                                                index=index).astype('int32'))
+                                      pd.Series(-1, index=index))
 
 
 # this deviates from the step in urbansim_defaults only in how it deals with
